# services/metadata/cmd/server/main.go
package main

// @title Flowix Metadata API
// @version 1.0
// @description Video metadata CRUD + internal status (Go chi). JWT Bearer required for write ops.
// @BasePath /
// @securityDefinitions.apikey BearerAuth
// @in header
// @name Authorization
// @description Type "Bearer {token}" for JWT
import (
	"context"
	"encoding/json"
	"log"
	"log/slog"
	"net/http"
	"net/url"
	"os"
	"strconv"
	"strings"

	_ "flowix/metadata/docs" // generated by swag init
	"flowix/metadata/internal/handler"
	mw "flowix/metadata/internal/middleware"
	"flowix/metadata/internal/repository"
	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/jackc/pgx/v5/pgxpool"
	"github.com/rs/zerolog"
	httpSwagger "github.com/swaggo/http-swagger"
)

func main() {
	port := os.Getenv("METADATA_PORT")
	if port == "" {
		port = "8002"
	}
	dbURL := os.Getenv("DATABASE_URL")
	if dbURL == "" {
		dbURL = "postgres://flowix:flowix@localhost:5432/flowix?sslmode=disable"
	}
	// pgx doesn't like sslmode, strip it for compat (same as auth)
	dbURL = stripSSLMode(dbURL)
	jwtSecret := os.Getenv("JWT_SECRET")
	if jwtSecret == "" {
		jwtSecret = "change-me-super-secret-jwt-key-32chars"
	}
	logger := zerolog.New(os.Stdout).With().Timestamp().Logger()

	pool, err := pgxpool.New(context.Background(), dbURL)
	if err != nil {
		log.Fatalf("pgxpool: %v", err)
	}
	defer pool.Close()
	if err := pool.Ping(context.Background()); err != nil {
		log.Fatalf("db ping: %v", err)
	}
	logger.Info().Str("port", port).Msg("metadata starting")

	repo := repository.NewVideoRepo(pool)
	vh := handler.NewVideoHandler(repo)

	r := chi.NewRouter()
	r.Use(middleware.Logger, middleware.Recoverer, middleware.RequestID)

	r.Get("/health", func(w http.ResponseWriter, r *http.Request) {
		writeJSON(w, r, http.StatusOK, map[string]string{"status": "ok", "service": "metadata"})
	})
	r.Get("/", func(w http.ResponseWriter, r *http.Request) {
		writeJSON(w, r, http.StatusOK, map[string]string{"service": "metadata", "version": "0.1"})
	})
	r.Get("/swagger/*", httpSwagger.WrapHandler)
	r.Get("/openapi.json", func(w http.ResponseWriter, r *http.Request) {
		http.Redirect(w, r, "/swagger/doc.json", http.StatusMovedPermanently)
	})
	// protected routes: need JWT
	r.Group(func(r chi.Router) {
		r.Use(mw.AuthMiddleware(jwtSecret))
		r.Post("/api/v1/videos", vh.Create)
		r.Patch("/api/v1/videos/{id}", vh.Update)
		r.Delete("/api/v1/videos/{id}", vh.Delete)
	})
	// public list/get
	r.Get("/api/v1/videos", vh.List)
	r.Get("/api/v1/videos/{id}", vh.Get)
	// internal (no auth)
	r.Patch("/internal/videos/{id}/status", vh.UpdateStatus)
	r.Get("/internal/videos/{id}", vh.GetInternal)

	log.Printf("metadata listening :%s", port)
	if err := http.ListenAndServe(":"+port, r); err != nil {
		log.Fatal(err)
	}
}

func stripSSLMode(s string) string {
	if !strings.Contains(s, "sslmode=") {
		return s
	}
	u, err := url.Parse(s)
	if err != nil {
		return s
	}
	q := u.Query()
	q.Del("sslmode")
	u.RawQuery = q.Encode()
	return u.String()
}

func writeJSON(w http.ResponseWriter, r *http.Request, status int, v any) {
	w.Header().Set("Content-Type", "application/json; charset=utf-8")
	w.WriteHeader(status)
	enc := json.NewEncoder(w)
	enc.SetEscapeHTML(false)
	if shouldPretty(r) {
		enc.SetIndent("", "  ")
	}
	if err := enc.Encode(v); err != nil {
		slog.Error("encode json failed", "error", err, "path", r.URL.Path)
	}
}

func shouldPretty(r *http.Request) bool {
	if v := r.URL.Query().Get("pretty"); v != "" {
		if b, err := strconv.ParseBool(v); err == nil {
			return b
		}
		return true
	}
	if v := os.Getenv("PRETTY_JSON"); v != "" {
		if b, err := strconv.ParseBool(v); err == nil {
			return b
		}
	}
	return false
}