# services/metadata/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Video metadata CRUD + internal status (Go chi). JWT Bearer required for write ops.",
        "title": "Flowix Metadata API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/videos": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "List videos",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Limit 1-100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Create video metadata",
                "parameters": [
                    {
                        "description": "Create",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/flowix_metadata_internal_model.CreateVideoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/flowix_metadata_internal_model.Video"
                        }
                    },
                    "400": {
                        "description": "invalid body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/videos/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Get video by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Video ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/flowix_metadata_internal_model.Video"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Delete video (owner only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Video ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Update video (owner only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Video ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/flowix_metadata_internal_model.UpdateVideoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/flowix_metadata_internal_model.Video"
                        }
                    },
                    "403": {
                        "description": "forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/internal/videos/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "internal"
                ],
                "summary": "Internal get video (no auth)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Video ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/flowix_metadata_internal_model.Video"
                        }
                    }
                }
            }
        },
        "/internal/videos/{id}/status": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "internal"
                ],
                "summary": "Internal update status (transcoder/upload)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Video ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/flowix_metadata_internal_model.UpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "flowix_metadata_internal_model.CreateVideoRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "flowix_metadata_internal_model.Rendition": {
            "type": "object",
            "properties": {
                "bitrate": {
                    "type": "integer"
                },
                "height": {
                    "type": "integer"
                },
                "quality": {
                    "type": "string"
                },
                "s3_key": {
                    "type": "string"
                },
                "video_id": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "flowix_metadata_internal_model.UpdateStatusRequest": {
            "type": "object",
            "properties": {
                "renditions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/flowix_metadata_internal_model.Rendition"
                    }
                },
                "status": {
                    "$ref": "#/definitions/flowix_metadata_internal_model.VideoStatus"
                }
            }
        },
        "flowix_metadata_internal_model.UpdateVideoRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "flowix_metadata_internal_model.Video": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "renditions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/flowix_metadata_internal_model.Rendition"
                    }
                },
                "status": {
                    "$ref": "#/definitions/flowix_metadata_internal_model.VideoStatus"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "flowix_metadata_internal_model.VideoStatus": {
            "type": "string",
            "enum": [
                "uploaded",
                "processing",
                "ready",
                "failed"
            ],
            "x-enum-varnames": [
                "StatusUploaded",
                "StatusProcessing",
                "StatusReady",
                "StatusFailed"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer {token}\" for JWT",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}